update int read_iob_frame() to unsigned long long read_iob_frame(), Fixes Issue 77
git-svn-id: https://hrpsys-base.googlecode.com/svn/trunk@641 a991ac11-fb38-5095-8c12-a1ddb0715245

+++ lib/io/iob.cpp
@@ -535,7 +535,7 @@ int read_torque_limit(int id, double *limit)
   return FALSE;
 }
 
-int read_iob_frame()
+unsigned long long read_iob_frame()
 {
     ++frame;
     if (frame == 5) frame = 0;
